refactor presentation

+++ src/main/java/lms/service/impl/PresentationServiceImpl.java
@@ -10,6 +10,7 @@ import lms.entities.Presentation;
 import lms.entities.Trash;
 import lms.enums.Type;
 import lms.exceptions.AlreadyExistsException;
+import lms.exceptions.BadRequestException;
 import lms.exceptions.NotFoundException;
 import lms.repository.LessonRepository;
 import lms.repository.PresentationRepository;
@@ -38,25 +39,27 @@ public class PresentationServiceImpl implements PresentationService {
     public SimpleResponse createPresentation(Long lessonId, PresentationRequest presentationRequest) {
         Lesson lesson = lessonRepository.findLessonById(lessonId)
                 .orElseThrow(() -> new NotFoundException("Урок с id: " + lessonId + " не существует!"));
-        boolean exists = presentationRepository.existsTitle(lesson.getId(), presentationRequest.getTitle());
-        if (exists) {
-            throw new AlreadyExistsException("Презентация с названием " + presentationRequest.getTitle() + " уже существует!");
-        }
-        boolean notNullTrashPresentations = presentationRepository.existsNotNullTrashPresentation(lesson.getId(), presentationRequest.getTitle());
-        if (notNullTrashPresentations)
-            throw new AlreadyExistsException("Презентация с названием " + presentationRequest.getTitle() + " уже есть в корзине!");
-        Presentation presentation = new Presentation();
-        presentation.setTitle(presentationRequest.getTitle());
-        presentation.setDescription(presentationRequest.getDescription());
-        presentation.setFile(presentationRequest.getFile());
-        presentation.setLesson(lesson);
-        presentationRepository.save(presentation);
-        lesson.getPresentations().add(presentation);
-        lessonRepository.save(lesson);
-        return SimpleResponse.builder()
-                .httpStatus(HttpStatus.OK)
-                .message("Успешно загружено!")
-                .build();
+        if (lesson.getTrash() == null) {
+            boolean exists = presentationRepository.existsTitle(lesson.getId(), presentationRequest.getTitle());
+            if (exists) {
+                throw new AlreadyExistsException("Презентация с названием " + presentationRequest.getTitle() + " уже существует!");
+            }
+            boolean notNullTrashPresentations = presentationRepository.existsNotNullTrashPresentation(lesson.getId(), presentationRequest.getTitle());
+            if (notNullTrashPresentations)
+                throw new AlreadyExistsException("Презентация с названием " + presentationRequest.getTitle() + " уже есть в корзине!");
+            Presentation presentation = new Presentation();
+            presentation.setTitle(presentationRequest.getTitle());
+            presentation.setDescription(presentationRequest.getDescription());
+            presentation.setFile(presentationRequest.getFile());
+            presentation.setLesson(lesson);
+            presentationRepository.save(presentation);
+            lesson.getPresentations().add(presentation);
+            lessonRepository.save(lesson);
+            return SimpleResponse.builder()
+                    .httpStatus(HttpStatus.OK)
+                    .message("Успешно загружено!")
+                    .build();
+        } else throw new BadRequestException("Урок может быть в корзине!");
     }
 
     @Override
@@ -65,54 +68,60 @@ public class PresentationServiceImpl implements PresentationService {
                                            EditPresentationRequest presentationRequest) {
         Presentation presentation = presentationRepository.findPresentationById(presentationId)
                 .orElseThrow(() -> new NotFoundException("Презентация с id:  " + presentationId + " не существует!"));
-        Lesson lesson = lessonRepository.findLessonByPresentationId(presentation.getId());
-        if (!presentation.getTitle().equals(presentationRequest.getTitle())) {
-            boolean exists = presentationRepository.existsTitle(lesson.getId(), presentationRequest.getTitle());
-            if (exists) {
-                throw new AlreadyExistsException("Презентация с названием " + presentationRequest.getTitle() + " уже существует!");
+        if (presentation.getTrash() == null) {
+            Lesson lesson = lessonRepository.findLessonByPresentationId(presentation.getId());
+            if (!presentation.getTitle().equals(presentationRequest.getTitle())) {
+                boolean exists = presentationRepository.existsTitle(lesson.getId(), presentationRequest.getTitle());
+                if (exists) {
+                    throw new AlreadyExistsException("Презентация с названием " + presentationRequest.getTitle() + " уже существует!");
+                }
             }
-        }
-        boolean notNullTrashPresentations = presentationRepository.existsNotNullTrashPresentation(lesson.getId(), presentationRequest.getTitle());
-        if (notNullTrashPresentations)
-            throw new AlreadyExistsException("Презентация с названием " + presentationRequest.getTitle() + " уже есть в корзине!");
-        presentation.setTitle(presentationRequest.getTitle());
-        presentation.setDescription(presentationRequest.getDescription());
-        presentation.setFile(presentationRequest.getFile());
-        presentationRepository.save(presentation);
-        return SimpleResponse.builder()
-                .httpStatus(HttpStatus.OK)
-                .message("Презентация успешно обновлена!")
-                .build();
+            boolean notNullTrashPresentations = presentationRepository.existsNotNullTrashPresentation(lesson.getId(), presentationRequest.getTitle());
+            if (notNullTrashPresentations)
+                throw new AlreadyExistsException("Презентация с названием " + presentationRequest.getTitle() + " уже есть в корзине!");
+            presentation.setTitle(presentationRequest.getTitle());
+            presentation.setDescription(presentationRequest.getDescription());
+            presentation.setFile(presentationRequest.getFile());
+            presentationRepository.save(presentation);
+            return SimpleResponse.builder()
+                    .httpStatus(HttpStatus.OK)
+                    .message("Презентация успешно обновлена!")
+                    .build();
+        } else throw new BadRequestException("Презентация может быть в корзине!");
     }
 
     @Override
     public PresentationResponse findById(Long presentationId) {
         Presentation presentation = presentationRepository.findPresentationById(presentationId).orElseThrow(() -> new NotFoundException("Презентация с id: " + presentationId + " не найден!"));
-        return PresentationResponse.builder()
-                .id(presentation.getId())
-                .title(presentation.getTitle())
-                .description(presentation.getDescription())
-                .file(presentation.getFile())
-                .build();
+        if (presentation.getTrash() == null) {
+            return PresentationResponse.builder()
+                    .id(presentation.getId())
+                    .title(presentation.getTitle())
+                    .description(presentation.getDescription())
+                    .file(presentation.getFile())
+                    .build();
+        } else throw new BadRequestException("Презентация может быть в корзине!");
     }
 
     @Override
     @Transactional
     public SimpleResponse deletePresentationById(Long presentationId) {
         Presentation presentation = presentationRepository.findPresentationById(presentationId).orElseThrow(() -> new NotFoundException("Презентация с id: " + presentationId + "не найден!"));
-        presentationRepository.deletePresentation(presentationId);
-        Trash trash = new Trash();
-        trash.setName(presentation.getTitle());
-        trash.setDateOfDelete(ZonedDateTime.now());
-        trash.setType(Type.PRESENTATION);
-        trash.setPresentation(presentation);
-        presentation.setTrash(trash);
-        trashRepository.save(trash);
-        presentationRepository.save(presentation);
-        return SimpleResponse.builder()
-                .httpStatus(HttpStatus.OK)
-                .message("Презентация успешно удален!")
-                .build();
+        if (presentation.getTrash() == null) {
+            presentationRepository.deletePresentation(presentationId);
+            Trash trash = new Trash();
+            trash.setName(presentation.getTitle());
+            trash.setDateOfDelete(ZonedDateTime.now());
+            trash.setType(Type.PRESENTATION);
+            trash.setPresentation(presentation);
+            presentation.setTrash(trash);
+            trashRepository.save(trash);
+            presentationRepository.save(presentation);
+            return SimpleResponse.builder()
+                    .httpStatus(HttpStatus.OK)
+                    .message("Презентация успешно удален!")
+                    .build();
+        }else throw new BadRequestException("Презентация может быть в корзине!");
     }
 
     @Override
